Create function to do plot for wood outcome
The main structure and the parameter settings are the same as in main_batch, but since we need all three forest configurations in the same plot for the average wood outcome. This file contains an additional for loop over the forest types. All the other calculations from the main were neglected for this file.